Implement data preprocessing & training
Update README's package install instructions for the tensorflow variant of galaxy-datasets since we now use its Dataset features

The notebook training and evaluating the model is using an explicitly smaller number of data samples as a proof of success. There have been time issues and errors using the full data.